Introduce ApplicationMetaData
So far we've kept every model in a hierarchy starting at
ApplicationModel. The models are actually not all the same though. Some
are the parsing result of JSON model files and some are secondary data
about those models or in the case of StaticFuntionReferences the result
of analysing the JavaScript modules.

We move the latter kind of models into a new model ApplicationMetaData
which then gets referenced by ApplicationModel in a one-to-one
relationship.

This also prevents application model from growing too much and keeps all
meta models nicely together in one container.