Implement resource writing and deleting
These basic operations have been missing so far. We add delete and write
methods on the resource loader and the resource itself and also an
isWritable method to check if a resource is writable.

Classpath and servlet context resources are never writable.

Note that the system will still mount base and local resources as file
resource roots if possible (exploded WAR for local resources, System
property exceed.library.source controlling that for base)