Refactor/Simplify scopes
The existing scope system based on three object types
property,object,list is too complicated, especially given the fact that
we can express all these value types as a property now.

Scoped Properties only
======================

We only have one type of scoped value that is a property based on a
extended domain property definition. Complex domain property values are
already used internally for the internal editors and are now used for
the scopes, too.

Scoped objects are now domain type properties like

    "context": {
        "properties": {
            "scopedFoo": {
                "type": "DomainType",
                "typeParam": "Foo",
                "defaultValue": "newObject()"
            }
        }
    }

For a domain object, the property "type" must be "DomainType" and the
"typeParam" must give the domain type name.

ScopeMetaModel
==============

So far we have been doing runtime lookups in the individual scopes to
work with scopes. This is both slow as well as prone to have scoped
values bleeding through were they shouldn't be visible just because a
value is in the context map.

Now we use a more structured approach that has some advantages:

 * quick one-level lookup of the valid scope declarations within an
   application execution context

 * startup validation of scope references (also potential merging
   of scope declarations)

After we have read in the intial models, we are already doing a post
processing phase which we reorganize a bit now and to which we add
scope analysis.

Application Scope Locations
===========================

So we generally have different types of application scopes with
different lifetimes and validity.

 * Application scope for general application configuration
 * Session scope for per-user application values
 * Process scope (versioned, one per process)
 * View scope (one per view)

We might add more later or provide more configuration options for these
(e.g. client-side scopes).

For each application execution location the scope declarations of all
related scopes together make up the total set of valid scoped property
declarations.

We create a scope meta model that contains references to all valid
declarations for all application locations.

This meta model we can use to both validate at startup and also to
lookup, convert etc at runtime.

Again we need to special handle the process context scope declaration
"current" which is now a "DomainType" typed property which has the type
of whatever the current object's type currently is.

Scope Expressions
=================

So far we have been using expressions like

    property('myName')

    object('myObj').num = 1

to access scoped properties which at that time I thought to be better as
it clearly shows the type of the scoped object.

But it also adds a lot of noise to the expressions and seems unnatural
compared to the very simple alternative of using identifiers.

I was put off my the fact that the identifiers were only used to refer
to the defined domain types like e.g.

	query(MyType)

In the end there will be no collision if we stick to the convention that
domain types / entities should begin with an uppercase letter and
properties / objects / lists with a lowercase letter.

	myName

        myObj.num = 1

are both much easier and less noisy to read. After the scope refactoring
we can easily lookup everything we need to know about a scope
declaration (validity, scope type, scoped property model etc) from the
new scope meta model.