Refactor RoutingTable
The JSON format for routing.json was much too complicated to input and
not really readable either. The format is good for internal resolution
purposes though.

We change the JSON format to be more user-friendly and also much easier
to implement an editor for.

Example from the test JSON file:

{
    "mappings": {
        "/": {
            "viewName": "Home"
        },
        "/detail/{id?}": {
            "viewName": "Detail"
        },
        "/scoped": {
            "processName": "scoped"
        },

        "/foo/special": {
            "processName": "specialFoo"
        },
        "/foo/{fooId}": {
            "processName": "normalFoo"
        },
        "/multi/{a}/{b?}/{c?}": {
            "viewName": "Multi"
        }
    }
}

The mappings property contains a map mapping location templates to
MappingNode JSON which contains either a viewName referencing a
non-process View to serve from that location or processName to serve a
process from that location.

Location templates can contain template variable parts. The variable
must span an entire URL part and can end with a ? marking the parameter
as optional.

We transform this structure to the original tree structure in the
mappings setter.

We add a test for parsing and vars handling.