Models: Implement Autoversioning
We need to deal with limited concurrency while editing modules which we
will solve by a simple MVCC system based on an identity guid and a
version guid.

Although the human editors refer to the models by name, we will track
the model identity with an autogenerated identity string (UUID for now)
this allows us easy handling of model renaming etc.

To handle concurrent editing, we add a second identity string called
version guid which is unique for every version of this model.

A user editing a model must give the version number he believes to be
updating. If the version number does not match, we know the resource
has been edited and we need to merge the differences.