Undo: Add done handlers, reactivate test
The undo helper already uses user-defined state functions as callbacks
to both encapsulate data management and reissue changes based on
undos/redos back into to the system.

These propagated changes then often lead to async processes being
triggered, which might be AJAX or just updating react state.

We might use promises here, but it seems easier to add another callback
which is the mechanism needed by the async operations so far anyway.
Since the usage is largely optional and only needed in a few cases, we
also have way less overhead in most cases.

Done method arguments
=====================

The optional done method argument is added to the methods

 * undoSvc.undo(done)
 * undoSvc.redo(done)
 * undoSvc.revert(done)
 * undoSvc.newState(state, done)
 * undoSvc.replaceState(state, done)
 * undoSvc.markSaved(done)

The state callback given to .create(initial, stateCallback) receives a
done parameter which has to be called after the state change has fully
propagated. It is guaranteed to be a (potentially noop) method at this
point.

Fix environment issues in event.js and reactivate the extended undo
test.