ModelComposition: Externalize rules
Having the model composition rules encoded just as code within the model
composition service is not ideal. This commit refactors the service to
externalize the rules into a map of rules that handles matching based on
simple rules. Either just a prefix rule

    /models/config.json => xcd.ApplicationConfig
    /models/domain/property/ => xcd.domain.DomainProperty

or a rule with a single * place holder making it a prefix/suffix rule

    /models/process/*/view => xcd.view.View

We configure these rules in the model config spring configuration.