Restructure views
So far, we had the view with a single component tree starting at a root
component. This commit extends that to a system with separated, flexible
layouts.

Instead of a root component, we have a "content" map inside the view
with named component trees. We create an external "LayoutModel" that has
a component tree with special <Content/> tags in it. These <Content/>
tags import the content areas from our views to build the final view.
The content component can have a name attribute to define a name for the
content to be imported, which defaults to "main".

The rendering process starts at the content "root", which by default is
just content component importing "main". "main" becomes our default
content name to model applications. All layouts are expected to use the
"main" slot as primary content area, though that is not a hard
requirement.

The layout can define its own context properties which will be merged
into every view the layout is applied to. (We maybe could also add means
to "pull up" these layout scope properties into the including process,
but so far it's only for view properties).