Model: Add documentation annotations
Add annotations to enable automatic generation of model documentation
from the model classes themselves. The javadoc is just too little,
doesn't provide the "JSON view" on things and also not the resource
directory structure.

In general we will use the svenson JSON configuration we are already
using to provide property names and type hints and information about
ignored properties etc. Then we add some more specialized annotations
to express some things in our model world.

The new annotations are as follows:

@Internal

Type/method is used internally and/or does not need to be modeled.

@DocumentedMapKey
Key name to use in the type description of map properties.

(format "Map keyName -> Type")

@DocumentedSubTypes

To be used in locations where the a property uses a super type of the
actual types involved. For example in Process, we have a state map where
the formal type is Map<String,ProcessState> and the type hint is also
annotated as such. ProcessState however is just an abstract base class
for the process state types we currently support. So instead of using
class scanning to find all implementations etc, we simplify things and
just list the types to document for that property in the order in which
we want them.

@DocumentedModelType

Used to change the documented type of of a property to make it more
logical / readable / clear from the documentation point of view.