Introduce ClientStateService/ClientStateProviders
Most of the work we do on the server side for exceed is preparing data
for consumption of the client. Apart from the component level data
providers, we have also been using something called a
RuntimeInfoProvider to provide us with all kinds of general information
about application state, authentication, process data and whatnot.

In preparation to moving over to redux for our state organization and
handling, we overhaul the system how we provide that state from the
server to the client.

We introduce a new interface now called ClientStateProvider and a
central service called ClientStateService. All client data is provided
by this service in one big data block created from the
ClientStateProvider instances used.

We distinguish between two kinds of providers: Providers that provide
directly in-view mutable data and data that mostly serves as a reference
for the current application configuration and state but is not subject
to change in the same way, only when the view changes and maybe not even
then.

Each provider defines a unique name and the mutable flag. The redux
state object is built out of all provider objects, the immutable objects
all moved into a "meta" block.

The provided blocks are very different in nature and lifecycle/scope.
Some never change, some are constant per application, some are constant
per user, some are constant per view model and so on. We let the
provider define a ClientStateScope enum value that defines the scope
they're in and use that to make the ClientStateService cache the JSON
blocks and only recreate those that are not fresh.

Provide method
==============

The provide method is the central service method of the provider which
is called whenever a fresh state slice is required.

    ClientData provide(
        HttpServletRequest request,
        RuntimeContext runtimeContext,
        ViewData viewData
    )

There are two implementations of ClientData: DefaultClientData which
uses a default JSONifier to convert its parameter into JSON and a
JSONData() implementation that takes already JSONified data as
parameter.