Fix potentially wrong file separator usage
So far we have only used on Util.path method to convert hard-coded
pathes with "/" to their system-dependent pendant.

This method already had been used wrongly for the opposite use-case in
one method, so it seems wise to rename the method to make it clear what
direction the conversion happens. We also need the opposite method which
converts a path with system-dependent file separators to one with
slashes for internal use.

We rename the old method to "toSystemPath" and introduce a mew method
"toSlashPath". Not the prettiest of names, but certainly disambiguated.