Component-Update: Fix combination with preview
Make the single component update work together with preview.

Move the updateComponent action producer into the view actions and also
make it use the internal fetch() function which automatically handles
the changedViewModels transmission. Use POST in any case.

Restructure DefaultRuntimeApplication a bit to parse the request
body into a Map on both preview and component update.

Draw the component update parameter from that map and move the
server-side handling behind the handling of the preview view generation
to actually do the component update with the preview view model.

Change the field name transmitting the current editor state to
"changedViewModels" in TransitionData and use that name in fetch.