ModelSchema: Introduce @ExceedPropertyType
So far, we have mostly been reusing the existing JSON annotations to
analyze the model classes and generate domain type definitions from
them. This works reasonably well in most cases but there are some edge
cases that didn't work yet.

This introduces a new method-level annotation @ExeedPropertyType that
directly defines the Exceed type and type param to use for a property.

This way, we can e.g. specify that a property which is formally
Map<String,DomainObject> is in fact a "Map" property with
"AppTranslation" param from the exceed perspective.