ModelLocationRules: Centralize
Move the constants and the default list of rules into the constructor of
model location rules. The rules are a result of the current
implementation state of the model classes and are very unlikely to
really meaningful change via spring configuration.

We keep an adder method around, just in case, but in the mean time, we
have everything in one place.