ExpressionRenderer: Fix superfluous brackets
Avoid superfluous brackets around expression by making the renderer
check whether the ASTExpression it is rendering has a parent or not.

If it has a parent, it is a subexpression nested somewhere in another
expression where the grammar recurses into itself.

Only in that case we actually might need the brackets and keep rendering
it. In every other case, we don't.

We adjust some test to no longer expect these outmost brackets.