Refactor expression values
The way we handle attributes in for the component model has spread like
wildfire through the code base and it is high time that we step in and
move the stuff to a more appropriate location and rename to to
"ExpressionValue".

It does everything the former "AttributeValue" did, that is parsing and
keeping the value of a expression for both a forced mode that always is
an expression and the kind of attributes we use in the components where
the expression can either be a string literal or is enclosed in "{ … }".

If the expression contains an error, that error will be stored in the
expression value object and can be requested from there. The expression
will be null in this case.