Cleanup property type handling
Don't just conclude from the name of the property type to its converter
name but always to the formal way of getting the formal PropertyType model
and looking up the registered converted via domain service.

Furthermore make sure that the PropertyType models don't reference the
class name of the converter but preconfigured instances. Currently all
in DefaultPropertyConverters.

This is in preparation of doing more elaborate things with the property
type definitions.