Expressions: Add Decimal, Refactor property chain
We prepare the expression language for decimal number support. We will
not be doing float or double numbers but instead either have integer
or a precise decimal type backed by bignumber.js on the js side and
BigDecimal on the java side.

We introduce several abstract classes and interfaces to make it easier
to work with the AST trees.

 * ArithmeticOperatorNode to distinguish the operators we will have to
   handle with Decimal

 * ResolvableNode implemeted by ASTFunction and ASTIdentifier to resolve
   language elements by name.

 * LiteralValueNode implemented by all nodes that represent a value and
   a corresponding type within the AST

Property Chain Refactoring
==========================

So far, we've been trying to handle object chaining with two different
container types ASTPropertyChain and ASTComputedPropertyChain, but the
grammar was not quite right and would disallow expressions like

    foo.values[0].bar

, switching from one type of chain to the other and back.

We change the grammar so that it now only knows ASTPropertyChain and
make that rule a chain of either .something or [something] links now
called ASTPropertyChainDot and ASTPropertyChainSquare which then contain
the actual chain payload nodes.

For ASTPropertyChainDot this is ASTIdentifier and ASTFunction children.

For ASTPropertyChainSquare this can be any expression.

We change the different systems and fix their tests to account for the
changes