Redo Action API
The first design of the action API for exceed was overly influenced by
the way it was done in OpenSAGA which is not what we really need here.

The new system is an annotation based system working on java methods
that provide both the base functionality we deliver with the exceed base
distribution, but also uses the same mechanism for application-side
extension and implementation of functionality.

See src/main/doc/action.md for detail