Refactor client code
Replace the react context based form handling with a redux state based
form handling.

We use a new state slice named "formState" to contain form data for the
current view.

This contains information about what forms, formblocks and form fields
are there, what there cursor paths relative to our scope graph are.

For textual inputs we need to be able to store arbitrary string values
for a field, even if they don't fit the underlying property type.

For example "2a" is clearly not a valid Integer value, but we
nevertheless need to store that value and present it to the user again
with an error note so they can change it to something valid.

We need to store information about the current form errors, too.

We implement an FKSelect widget to select the targets of foreign keys
from a popup.

We implement a Dialog component that allows us to model modal dialogs
and control from with a DialogState value.

We implement basic transition confirmation.