Startup: Move hub init into ace-loader
The initialization of the websocket takes some time and so far, we only
use it for the editor, so we move the initializtion into the ace-loader
module and trigger it at the same time we load the ACE api.

This both improves normal page startup time and also hides the init time
in the code editor initialization time.