Expand Exceed docs
Add an markdown documentation for the models and expressions

Document and refactor the models. Some models where awkwardly located
both for our purposes here in docs generation but also generally.

We move EnumType and StateMachine to the root of the domain hierarchy,
to have both under "domain" and to distinguish those definitions from
the definitions of the actual domain types / query domain types.

Add a src/main/doc/theme folder containing an independent copy of
bootstrap, our theme and the doc-template.html.