Introduce authSchema
So far, we've been using the "public" schema for auth purposes which was
clearly wrong. We now implement an auth scheme that works out of the box
with a single app and supports multiple independent apps as well as apps
with a joined user pool.

We add an "authSchema" property in application config with defaults to
the normal "schema" by default. We use the authSchema whenever we access
the AppUser domain type.

In the user details service we query all active applications for the
username and return the first returned, remembering the authSchema as
special role "SCHEMA_mySchema".

In the application router, we check for that role in the AppAuthentication
, if it is not present a security error is thrown.

Applications can define other schemas as authSchemas to share their user
pool. It is ensured that no user can access an application unless he
authenticated against the AppUser table in the authSchema.