Refactor DB schema handling, configuration
This commit cleans up the way exceed handles data storage with the
default JOOQDomainOperations.

We have been using dedicated application schemas for some time now but
still had remnants of earlier development phases that were a mix
of model-generated and hand-created database tables in the "public"
schema. These tables contained data that was either used for spring
security purposes (persistent tokens for the "Remember me" feature)
parts application data inconveniently located at that time.

The exceed services were designed to support multi-tenant capabilities
but the startup and configuration did not yet allow to configure it.

Schema usage
------------

We extend the config.json configuration to allow the definition of an
"authSchema" in addition to the already existing "schema". This
authSchema is used for all authentication related domain types and can
be located in another schema to share user groups with another
application.

The ROLE_ADMIN role becomes a global role for all running applications.
We try to authenticate a user against all existing application
authSchema now and remember which schema authenticated the user as a
special role (e.g. SCHEMA_myApp).

We make sure that the user has that schema role when accessing an
application (or is admin).

In the future we have to tightly control who can assign those SCHEMA_xxx
roles to prevent security escalation in multi-tenancy scenarios etc.

exceed-config.json
------------------

So far, we had our basic setup configuration in a property file
/WEB-INF/cfg/exceed-app.properties.

We now move that to a JSON config file /WEB-INF/cfg/exceed-config.json

```json
{
    "env" : {
        "spring.profiles.active" : "development",

        "deactivated.apps" : [
        ]
    },

    "apps" : [
        {
            "name" : "shipping",
            "extensions" : [
                "shipping-app"
            ]
        },
        {
            "name" : "crud",
            "extensions" : [
                "crud-example"
            ]
        }
    ]
}
```

The apps array property contains an array of app-state object (what used
to be the app_state table) which defined what applications there are,
what there name is, where the extensions for it are etc.

The "env" property contains a map of Spring env properties. Since you
can't comment out apps in the JSON format (stupid Crockford), you can
move the definitions e.g. a "deactivated.apps" array property  if you
want to deactivate them totally.

The first app in this definition is considered the default app by
default. You can change that by defining a "defaultApp" property naming
a different active app.

The default app is used to:

 * Define which app the root URI / is redirected to
 * What style to use for the login view
 * in which authSchema to store persistent tokens for "Remember me"

Appliction and User config
--------------------------

So far, we have stored our "applicationContext" in a JSON blob inside
the app_state table which is now gone.

Instead we move the "applicationContext" and a new "userContext" for
user-specific persistent properties to two new partial base domain
types.

The base definition now defines two domain types called "AppConfig" and
"AppUserConfig" as a skeleton definition with only an id and another
application/user identifying field.

That skeleton definition then gets enhanced during application startup
to contain columns for each context property in "applicationContext" and
"userContext".

We keep automatically synching those contexts back to the storage so
that the user never has to deal with them, they can just declare and
access the context properties.

For now we support a limited number of non-complex property types for
these context properties. We might add support for all property types
as jsonb/JSON-in-varchar columns later.

AppLogin
--------

We introduce another new base domain type "AppLogin" which stores the
information formerly kept in a JDBC accessed table "persistent_logins".
It stores the token information for Spring security persistent tokens
used for the Spring Security "Remember me" feature.

This domain type is only used for the default application storing the
tokens for all users of all apps.

Translations
------------

Translations also were taken only from "public" so far, we implement
a new ExceedAppTranslationProvider to actually read from the correct
application schema.

Unique Constraints
------------------

Exceed supports unique constraints on single columns now. Just add
a "unique" : true property to the domain type property to make it unique
and automatically have a DB unique constraint defined for it.

JOOQ Generation for base
------------------------

Since we now moved everything out of "public" into the appropriate
application schema/authSchema, we have nothing to generate classes for
in base. We keep the old generated classes around for now as we still
use them internally as typed containers.