Add componentConfig for base template
We already provided some convenience placeholders for base template
customization purposes but provided only a Java interface with which
those placeholders can be filled.

While this is certainly a good feature to have for "HTML-blob"
integration scenarios, those are an ugly escape hatch and not something
one should use for simple stuff.

We already added ComponentConfig as a general purpose container for
model-driven component configuration to which we now add a
BaseTemplateConfig. We make it a DynamicProperties instance to be able
to accept any value (templates can be replaced) but which also allows
us to create formal setters accessing the same dynanmic properties for
documentation purposes.