Implement model merger
Create a separate ModelMerger class that was as an annotation driven
merger to

 * Merge models at the same location from different extensions
 * Merge the top level models into the application model hierarchy

We use a new model annotation @MergeStrategy(MergeType.REPLACE|DEEP) to
control the merging behavior for the models.

REPLACE replaces top level models in lower extensions with those in the
higher priority extensions

DEPP does a deep object merge on the top level model.

Some property types are always REPLACEd:

 * Primitives
 * Numbers
 * Strings
 * AppResource instances

Inside a model container that marked with
@MergeStratefy(MergeType.DEEP), you can influence the merge type for the
inner properties by either putting an @MergeStrategy annotation on one
of the property setters or getters or directly on the contained types.

So is for example RoutingTable merging deep, and so is the `mappings` map
inside it, but the Mapping class itself is marked
@MergeStrategy(MergeType.REPLACE) to be replaced as a whole.

The merging behavior seems to work ok so far for the current use cases,
we might fine-tune the behavior or add new merge strategies as needed.

We also need to extend the generated documentation to document merge
behavior.