Build: Check in js build productsi (Experimental)
We remove the build products from .gitignore to add them to the git
repo in their dev-version. This allows users to just checkout the exceed
and exceed-test repos and run it in Intellij without needing to run a
js build, either normal or via maven on their system.

This goes somewhat against conventional wisdom of what to keep in git,
but I think it might work here. Don't bother with merge conflicts. Make
sure that you have the current version of the code and replace all of
these files, `git checkout --ours` if necessary.

If this causes nasty results we will have to remove it again.