Switch to react-16.1.0-rc
We add the dependency "es6-shim" for its Map/Set implementation needed
by react 16. Up to now we did get away with not having it on the server
since we didn't really used the "renderStatic" property-convert method
that pulls it in on the server.

Add es6-shim to the server bundle so it still works.