ResourceInjector: Add conditional injection
We add the ability to conditionally inject resource files into meta
classes.

We add an interface ResourceInjectorPredicate that controls the
conditional injection of a resource into a meta class.

```java

    @InjectResource(
        value = "/resource/js/myscript.js",
        predicate="myPredicate"
    )
```

will inject the resource /resource/js/myscript.js if the
ResourceInjectorPredicate implementation with the spring bean name
"myPredicate" returns true.