Implement server-side rendering
Server-side rendering can be a very attractive option to provide
additional page-load speed and SEO friendliness to a react js
application.

This commit implements server-rendering for exceed which thankfully is
very easy due to the general architecture.

We introduce a new entry point ./server-render.js that provides react js
renderToString functionality. It draws in all of the react
infrastructure, redux and all the components in the server bundle
though. Which is used to create one JsEnvironment per user thread. So it
comes with a bit of memory overhead.

Configuration
-------------

We add two componentConfig options "serverRendering" and
"serverRenderingStrategy".

"serverRendering" is the general mode and can be one of:

 * DISABLED
 * ALWAYS
 * PER_STRATEGY

"serverRenderingStrategy" contains the spring bean name of a
ServerRenderStrategy implementation that controls whether a
request is rendered with or without pre-rendering.

Js-Build Bundles
----------------

We now have 3 server side js bundles.

 * "exceed-srvcommon.js" contains the common dependencies for
   all server entry points.

 * "exceed-server.js" still contains the server-side exceed code

 * "exceed-srvrender" contains all the extra server-rendering code and
   is only loaded when "serverRendering" is not DISABLED