Introduce ClientOnly HOC
We recently added server rendering, but we still had some cases in which
server rendered markup was not the same as the client hydration.

For example for the SVG charts, we scale them relative to the font-sizes
we detect on the client. This detection of course is meaningless on the
server and would just lead to markup changes in terms of scale factors
and sizes.

We introduce a high order component ClientOnly that wraps another
Component and makes it render nothing or a placeholder when we do server
rendering or the initial client hydration rendering.

Then on compoment mount, the HOC rerenders the component once again in
its final form.

We add that that HOC to BarChart and RadarChart for the SVG measurement
issues and to our InPageEditor