Remove bundles
And so ends the short history of putting the generated js bundles into
the git repo. It is just too annoying and the benefits now seem to be
not worth it.

If you want a exceed development setup, you need to run the maven build
once or install nodejs and yarn locally and run the watch mode.