Move @EnableAutoConfiguration to library project
Since the introduction to atomikos data sources we can't just add an
@EnableAutoConfiguration to our configuration but need to exclude
Spring's data source configuration.

To spare app developers from doing that over and over again we move the
@EnableAutoConfiguration into the base library file supposed to be
imported by the application.

If the application author wants to remove auto configuration they have
write a replacement for
de.quinscape.exceed.runtime.ExceedApplicationConfiguration and include
that as base config.