POM: Add OWASP security check
Add build plugin that automatic checks our maven dependencies and
transitive dependencies for CVE vulnerabilities

And we already found potential vulnerabilities, so we update netty
to a newer version.

Suppression of CVE-2017-8806
----------------------------

Unfortunately, no update to postgresql can prevent us being flagged for
CVE-2017-8806, which is doubly unfortunate as the issue is about
locally installed postgresql administration scripts and has absolutely
nothing to do with the postgresql.jar, not matter what version.

So we add a generall suppression file to the misc folder and suppress
CVE-2017-8806 for the postgresql dependency.